+++ orchestrator/src/client/components/index.ts
@@ -1,3 +1,4 @@
+export { FitAssessmentContent } from "./FitAssessmentContent";
 export { JobBriefPane } from "./JobBriefPane";
 export { JobDescriptionPanel } from "./JobDescriptionPanel";
 export { JobHeader } from "./JobHeader";
@@ -6,6 +7,7 @@ export { ManualImportSheet } from "./ManualImportSheet";
 export { OpenJobListingButton } from "./OpenJobListingButton";
 export { PipelineProgress } from "./PipelineProgress";
 export { ScoreIndicator } from "./ScoreIndicator";
+export { isAwaitingAiScore, ScoreRing } from "./ScoreRing";
 export { Stats } from "./Stats";
 export { StatusBadge } from "./StatusBadge";
 export { TailoredSummary } from "./TailoredSummary";
